requirements.txt
pydoll-python==2.15.1
aiohttp>=3.9.5,<4.0.0
asyncio==3.4.3
nest-asyncio==1.6.0
psutil==5.9.8
pyvirtualdisplay==3.0
websockets>=14,<15
aiohttp